+++ app/pages/admin/stations/stations.html.js
@@ -1,6 +1,0 @@
-/* View Specific JS */
-F1.deferred.push(function initPage() {
-
-  console.log( '[Salon Stations Page] Says Hi!' );
-
-});