+++ app/themes/salon/admin__header.html.php
@@ -1,3 +1,0 @@
-<header class="main-header">
-  <h2><?=$view->title?></h2>
-</header>