+++ app/themes/salon/comp__dateselect.php
@@ -2,7 +2,7 @@
 <div id="date-select-modal" class="modal">
   <div class="modal-inner">
     <header>
-      <h3>Date Select</h3>
+      <h3 class="modal-title">Date Select</h3>
       <a class="modal-close">Close X</a>
     </header>
     <div class="modal-body"><span id="date-nav-calendar">Calendar goes here...</span></div>