+++ app/themes/salon/comp__slideswitch.php
@@ -1,0 +1,7 @@
+<label class="switch[:class]" for="[:id]">
+  <span class="switch__title">[:title]</span>
+  <div class="switch__control"><input id="[:id]" name="[:id]" type="checkbox"
+      onchange="this.parentElement.parentElement.classList.toggle('on')[:onchange]">
+    <i></i>
+  </div>
+</label>