+++ app/themes/salon/doc__foot.php
@@ -2,7 +2,9 @@
   <script src="<?=$scriptFileRef?>"></script>
   <?php endforeach; ?>
 
-  <script type="module"><?php include $view->getInlineScriptFile(); ?></script>
+  <script type="module">
+    <?php include $view->getInlineScriptFile(); ?>
+  </script>
 
   <script src="js/main.js" type="module"></script>