+++ app/pages/admin/stations/changelog.md
@@ -2,3 +2,15 @@
 
 ## 09 Jan 2023 - INIT - Ver 1.0.0
   - Initial commit
+
+## 18 Jan 2023 - FT - Ver 1.1.0
+  - Build to 80% complete.
+  - Add `station__form.php`
+  - Add `station__row.php`
+  - Add `models/station.php`
+  - Add `stations.vm.php`  
+  - Add custom styles
+  - Add custom script
+  - Add comp__slideswitch
+  - Add list-row `<template>` section
+  - Add controller logic